requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
httpx==0.28.1
openai==1.109.1
pydantic>=2.0.0
streamlit>=1.28.0
aiofiles>=23.0.0
